Sine Wave Single LED
This code is created to talk with the Max patch by listening to the
Serial Port and getting raw data. The brightness of the LED changes
according to the base value sent from Max. *** You have to unplug the
USB and re-open the Max after any changes for the new code to work.